% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaymod-package.R
\docType{package}
\name{decaymod-package}
\alias{decaymod-package}
\alias{decaymod}
\title{The 'decaymod' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. https://mc-stan.org
}
