% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_CV_RE_stan.R
\name{ne_CV_RE_stan}
\alias{ne_CV_RE_stan}
\title{Bayesian negative exponential function with Stan,
including cross validation, with random effects}
\usage{
ne_CV_RE_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian negative exponential function with Stan,
including cross validation, with random effects
}
