% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_neg_exp-data.R
\docType{data}
\name{sim_neg_exp}
\alias{sim_neg_exp}
\title{Simulated negative exponential data}
\format{A dataframe}
\source{
Saras Windecker
}
\usage{
data(sim_neg_exp)
}
\description{
Simulated negative exponential data
}
\examples{
data(sim_neg_exp)
head(sim_neg_exp)
}
\keyword{datasets}
