% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_weibull-data.R
\docType{data}
\name{sim_weibull}
\alias{sim_weibull}
\title{Simulated weibull data}
\format{A dataframe.}
\source{
Saras Windecker
}
\usage{
data(sim_weibull)
}
\description{
Simulated weibull data
}
\examples{
data(sim_weibull)
head(sim_weibull)
}
\keyword{datasets}
