% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_CV_noRE_stan.R
\name{w_CV_noRE_stan}
\alias{w_CV_noRE_stan}
\title{Bayesian weibull function with Stan,
including cross validation, without random effects}
\usage{
w_CV_noRE_stan(
  mT,
  mT_test,
  m0,
  m0_test,
  time,
  time_test,
  sp,
  J,
  X_alpha,
  X_alpha_test,
  X_beta,
  X_beta_test,
  ...
)
}
\arguments{
\item{mT}{Mass remaining of training data}

\item{mT_test}{Mass remaining of test data}

\item{m0}{Initial mass of training data}

\item{m0_test}{Initial mass of test data}

\item{time}{Vector of time values for training data}

\item{time_test}{Vector of time values for test data}

\item{sp}{Numeric group identifier}

\item{J}{Number of groups}

\item{X_alpha}{Model matrix for alpha for training data}

\item{X_alpha_test}{model matrix for alpha for test data}

\item{X_beta}{Model matrix for beta for training data}

\item{X_beta_test}{model matrix for beta for test data}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian weibull function with Stan,
including cross validation, without random effects
}
