% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_noCV_noRE_stan.R
\name{w_noCV_noRE_stan}
\alias{w_noCV_noRE_stan}
\title{Bayesian weibull function with Stan,
without cross validation, without random effects}
\usage{
w_noCV_noRE_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian weibull function with Stan,
without cross validation, without random effects
}
