% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_noCV_RE_stan.R
\name{w_noCV_RE_stan}
\alias{w_noCV_RE_stan}
\title{Bayesian weibull function with Stan,
without cross validation, with random effects}
\usage{
w_noCV_RE_stan(
  mT,
  m0,
  time,
  sp,
  J,
  X_alpha,
  X_beta,
  m0_sim,
  time_sim,
  sp_sim,
  ...
)
}
\arguments{
\item{mT}{Log mass remaining of training data}

\item{m0}{Log initial mass of training data}

\item{time}{Vector of time values for training data}

\item{sp}{Numeric group identifier}

\item{J}{Number of groups}

\item{X_alpha}{Model matrix for alpha for training data}

\item{X_beta}{Model matrix for beta for training data}

\item{m0_sim}{Simulated log initial mass data}

\item{time_sim}{Simulated time data}

\item{sp_sim}{Numeric group identifier for sim dataset}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian weibull function with Stan,
without cross validation, with random effects
}
