% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decaymod-package.R
\docType{package}
\name{decaymod-package}
\alias{decaymod-package}
\alias{decaymod}
\title{The 'decaymod' package.}
\description{
This package contains the compiled stan code for Weibull and negative exponential decay models with options for random effects, fixed effects on each decay parameter, and cross validation.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
