% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_CV_RE_stan.R
\name{ne_CV_RE_stan}
\alias{ne_CV_RE_stan}
\title{Bayesian negative exponential function with Stan,
including cross validation, with random effects}
\usage{
ne_CV_RE_stan(mT, mT_test, m0, m0_test, time, time_test, sp, J, X, X_test, ...)
}
\arguments{
\item{mT}{Mass remaining of training data}

\item{mT_test}{Mass remaining of test data}

\item{m0}{Initial mass of training data}

\item{m0_test}{Initial mass of test data}

\item{time}{Vector of time values for training data}

\item{time_test}{Vector of time values for test data}

\item{sp}{Numeric group identifier}

\item{J}{Number of groups}

\item{X}{Model matrix for training data}

\item{X_test}{model matrix for test data}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian negative exponential function with Stan,
including cross validation, with random effects
}
