% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_noCV_noRE_stan.R
\name{ne_noCV_noRE_stan}
\alias{ne_noCV_noRE_stan}
\title{Bayesian negative exponential function with Stan,
without cross validation, without random effects}
\usage{
ne_noCV_noRE_stan(mT, m0, time, sp, J, X, ...)
}
\arguments{
\item{mT}{Mass remaining of training data}

\item{m0}{Initial mass of training data}

\item{time}{Vector of time values for training data}

\item{sp}{Numeric group identifier}

\item{J}{Number of groups}

\item{X}{Model matrix for training data}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian negative exponential function with Stan,
without cross validation, without random effects
}
