% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_noCV_noRE_stan.R
\name{w_noCV_noRE_stan}
\alias{w_noCV_noRE_stan}
\title{Bayesian weibull function with Stan,
without cross validation, without random effects}
\usage{
w_noCV_noRE_stan(mT, m0, time, sp, J, X_alpha, X_beta, ...)
}
\arguments{
\item{mT}{Mass remaining of training data}

\item{m0}{Initial mass of training data}

\item{time}{Vector of time values for training data}

\item{sp}{Numeric group identifier}

\item{J}{Number of groups}

\item{X_alpha}{Model matrix for alpha for training data}

\item{X_beta}{Model matrix for beta for training data}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian weibull function with Stan,
without cross validation, without random effects
}
